### 2025-04-06 #8007 @editorialbot
[PRE REVIEW]: ActiSleep Tracker: a Python-based dashboard for adjusting automatic sleep predictions of actigraphy data
**Submitting author:** @nathaliaesper (<a href="http://orcid.org/0000-0003-3218-9226">Nathalia Bianchini Esper</a>)
**Repository:** https://github.com/childmindresearch/actisleep-tracker
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
