### 2025-05-06 #8007 @juliaferraioli
Hi @cbrnr, @richrobe, @TomDonoghue would you be interested in reviewing this paper for JOSS?
