### 2025-05-19 d8375c8
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/astral-sh/ruff-pre-commit: v0.6.8 → v0.11.10](https://github.com/astral-sh/ruff-pre-commit/compare/v0.6.8...v0.11.10)
- [github.com/pre-commit/pre-commit-hooks: v4.5.0 → v5.0.0](https://github.com/pre-commit/pre-commit-hooks/compare/v4.5.0...v5.0.0)
