### 2025-05-22 #8181 @cbrnr
I'd like to begin with some general comments that should be addressed before I proceed with the software review.

1. It would be nice if uv was supported instead of (or at least in addition to) poetry, since uv has gained a lot of traction in the Python community.
2. Would it be possible to extend support for Python 3.13, or is there a specific reason why only Python 3.11 is supported?
3. It is always difficult to assess if all authors have made major contributions to the software, because the Git history might be incomplete (for example, if there was significant development before the initial commit). Judging from the available Git history, the only major contributor is @reindervdw-cmi (the second author), which means that currently I cannot tick the corresponding checkbox in the review checklist. @nathaliaesper could you provide an overview of author contributions, specifically addressing if you made major contributions to the software and the role of the other authors?
4. The documentation website is not available, but it is linked from a badge in README.md (https://childmindresearch.github.io/).
5. Although installing the app is no problem, I need an actual data set to run the app. Would it be possible to provide some toy data (or maybe link to some publicly accessible data set) so that I can run the app? I think it would actually be a good idea to include such an example in the repository so that users can quickly try out the app without having to bring their own data.
