### 2025-05-28 #8181 @nathaliaesper
Hi @cbrnr
We appreciate your comments and reviews.

**1. It would be nice if uv was supported instead of (or at least in addition to) poetry, since uv has gained a lot of traction in the Python community.**
Thank you for the suggestion regarding uv. Since this release is focused on the end-user, we will consider support for uv in addition to poetry.

**2. Would it be possible to extend support for Python 3.13, or is there a specific reason why only Python 3.11 is supported?**
The recommended way for end-users to run the application is via the Docker image, which ensures a controlled and reproducible environment. Since this release is focused on the end-user, we will consider upgrading/supporting newer versions of Python.

**3. It is always difficult to assess if all authors have made major contributions to the software, because the Git history might be incomplete (for example, if there was significant development before the initial commit). Judging from the available Git history, the only major contributor is @reindervdw-cmi (the second author), which means that currently I cannot tick the corresponding checkbox in the review checklist. @nathaliaesper could you provide an overview of author contributions, specifically addressing if you made major contributions to the software and the role of the other authors?**
Thank you for raising this important point regarding author contributions. You are right that Git history alone can sometimes be an incomplete record, especially when a project has evolved from earlier work or has undergone significant restructuring. This application had grown iteratively, starting from a much smaller scope. Over time, new features and architectural improvements were introduced, leading to the current refactored version (https://github.com/childmindresearch/actisleep-tracker). The current codebase was built on a previous project I developed, which served as the foundational prototype (https://github.com/nathaliaesper/Actigraphy_APP).
Here is the role of all the authors:
@nathaliaesper: designing and implementing the original version of the application; architecting the software structure that guided the current design; implementing core functionalities that persist in the current version; testing the application; project leader
@reindervdw-cmi: code optimization; extensive unit testing; addition of features that brought the application to its current level of robustness and usability; led the transition from the prototype to a production-ready version.
Larissa Hunt: testing and validation, feature suggestions, project design
Céline Vetter: testing and validation, feature suggestions
Andrew Leroux: testing and validation, feature suggestions
Vadim Zipunnijov: testing and validation, feature suggestions
John Vito d’Antonio-Bertagnolli: project management
Michelle Freund: feature suggestions, project design, project management
Kathleen Merikangas: project design, conception, and project management
Michael Milham: project design, feature suggestions
Gregory Kiar: project design, management, and validation, led the transition from the prototype to a production-ready version
Alexandre R. Franco: project design, management, and validation, feature suggestions.

**4. The documentation website is not available, but it is linked from a badge in README.md (https://childmindresearch.github.io/).**
Thank you for pointing out that the documentation link in the README.md is incorrect. We are currently expanding the README.md file to include a usage example, demonstrating how to use the application once it is running.

**5. Although installing the app is no problem, I need an actual data set to run the app. Would it be possible to provide some toy data (or maybe link to some publicly accessible data set) so that I can run the app? I think it would actually be a good idea to include such an example in the repository so that users can quickly try out the app without having to bring their own data.**
We have included an example of data in the paper/data folder. You will find:

- raw data collected with an ActiGraph watch (paper/data/raw)
- raw data processed on GGIR  (paper/data/ggir_outputs)
- script used to process the raw data on GGIR (paper/data/scripts)

**6. I think the Docker instructions require that the working directory is set to the app's root directory; this is missing in the instructions and should be added.**
We are working on changing the instructions to clarify that the ${LOCAL_DATA_DIR} should be the path where the processed data is located.

**7. I watched the YouTube video, and I think this is really helpful! It would be nice if there were some explanations, ideally via audio or at least with some on-screen text. Currently, the audio track only contains ambient noise, so I recommend to either completely remove the audio or record a new audio track with comments.**
We are updating the instructions video uploaded on YouTube to include clear instructions on how to set up and use the application.
