+++ README.md
@@ -1,28 +1,111 @@
-# App Actigraphy
+# ActiSleep Tracker
 
 [![Build](https://github.com/childmindresearch/app-actigraphy/actions/workflows/test.yaml/badge.svg?branch=main)](https://github.com/childmindresearch/app-actigraphy/actions/workflows/test.yaml?query=branch%3Amain)
 [![Ruff](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/astral-sh/ruff/main/assets/badge/v2.json)](https://github.com/astral-sh/ruff)
 ![stability-wip](https://img.shields.io/badge/stability-work_in_progress-lightgrey.svg)
 [![L-GPL License](https://img.shields.io/badge/license-L--GPL-blue.svg)](https://github.com/childmindresearch/app-actigraphy/blob/main/LICENSE)
-[![pages](https://img.shields.io/badge/api-docs-blue)](https://childmindresearch.github.io/app-actigraphy)
 
-This webapp is an application designed for annotating sleep data. This repository contains the source code and related files for the application.
+ActiSleep Tracker is an application designed to run after processing actigraphy data with the [GGIR package](https://wadpac.github.io/GGIR/) and allows users to annotate sleep data. This repository contains the source code and related files for the application.
 
 ## Getting Started
 
-The app may be installed either through Docker (recommended for users) or Poetry (recommended for developers), see the instructions for each below. Whichever method you use to launch the app, the app will be available at http://localhost:8051.
+The application may be installed either through Docker (recommended for users) or Poetry (recommended for developers). Whichever method you use to launch the app, the app will be available at http://localhost:8051.
 
-### Running the App through Docker
+### Running ActiSleep Tracker through Docker
 
-1. Ensure you have Docker installed.
-2. Run the Docker image:
+1. Install Docker: Ensure you have Docker installed on your system. [Get Docker](https://docs.docker.com/get-started/get-docker/)
+2. Pull the Docker image:
+   ```bash
+   docker pull cmidair/actigraphy:main
+   ```
+3. Run the Docker image with your data:
    ```bash
    docker run \
       -p 127.0.0.1:8051:8051 \
-      --volume ${LOCAL_DATA_DIR}:/data \
-      --volume `pwd`/assets:/app/assets \
+      --volume /local/path/to/data:/data \
       cmidair/actigraphy:main
    ```
+Replace `/local/path/to/data` with the path to your input data directory (i.e., GGIR's output folder).
+
+## How to use ActiSleep Tracker?
+
+All the steps below will consider that you cloned this GitHub repository inside your Documents folder.
+There is an examplar raw data under `paper/data/raw` and a processed version using GGIR 3.2-6 under `paper/data/ggir_outputs`.
+
+After cloning this repository or processing your actigraphy data with GGIR, you should have a folder structure similar to:
+
+```
+├── 📁 ggir_outputs
+│ └── 📁 output_sub001
+│     ├── 📄 config.csv
+│     ├── 📁 meta
+│     └── 📁 results
+├── 📁 raw
+│ └── 📄 sub001.gt3x
+└── 📁 scripts
+    └── 📄 run_ggir.R
+```
+
+where `outputs` contains all files generated by GGIR; `raw` is the raw actigraphy data; and `scripts` contains all the scripts used in this project.
+
+Run ActiSleep Tracker through Docker by typing:
+
+```bash
+   docker run \
+   -p 127.0.0.1:8051:8051 \
+   --volume /Users/nathalia.esper/Documents/actisleep-tracker/paper/data/ggir_outputs:/data \
+   cmidair/actigraphy:main
+```
[…]
