### 2025-06-02 5cb5d94
Adding CONTRIBUTING.md file
