### 2025-06-03 4c1d78a
Adding CONTRIBUTING.md file (#97)
* Adding CONTRIBUTING.md file

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
