### 2025-06-06 #8181 @richrobe
Hi @nathaliaesper, first of all, thank you for this nice package! 😊

Besides my own, rather minor, comments (see below), I agree with the comments of @cbrnr, especially regarding the first point.

We have also been using `poetry` to manage our Python projects for a while, but we have fully moved to `uv` over the past couple of months due to the overwhelming advantages uv has over poetry. Migrating from `poetry` to `uv` is super easy by simply calling `uvx migrate-to-uv` (https://github.com/mkniewallner/migrate-to-uv). I also did it with this package, because I'm currently using Python 3.10 and therefore couldn't install the package that straightforwardly using `poetry`. Migrating the project to `uv` went smoothly because `uv` not only manages the virtual environment, but also the correct Python interpreter.

After migrating to `uv`, I could easily start the ActiTracker package and run some tests. Besides the comments already mentioned by @cbrnr, I had some comments on the functionality:
* "Add/Remove Slider" => maybe provide more explanation on what this does, i.e., that the sliders are used for adding another sleep period
* Double-clicking to reset the original view causes a misalignment between the plot and the sliders. Maybe this can be fixed by redrawing the plot after resetting the view? Adjusting the sliders after resetting the view re-aligns the plot.
* What is the y-axis of the plot? And since I assume that the units for x-axis angle and arm movement are not the same, I would suggest splitting the y-axis into two axes (one left of the plot and one right of the plot) to make it clearer what is plotted on which axis.

Besides that, I would suggest double-checking the paper for typos (e.g., l. 71: "Figure Figure 3"), and check the references to also include DOIs, etc.
