+++ pyproject.toml
@@ -1,49 +1,44 @@
-[tool.poetry]
-name = "actigraphy"
+[project]
+name = "actisleep-tracker"
 version = "0.1.0"
-description = "A beautiful description."
-authors = ["Reinder Vos de Wael <reinder.vosdewael@childmind.org>"]
-license = "LGPL-2.1"
+description = "Add your description here"
 readme = "README.md"
-packages = [{include = "actigraphy", from = "src"}]
+requires-python = "~=3.11"
+dependencies = [
+    "dash~=2.15",
+    "dash-bootstrap-components>=1.7.1",
+    "dash-daq>=0.6.0",
+    "numpy>=2.3.0",
+    "pandas>=2.3.0",
+    "plotly>=6.1.2",
+    "polars>=1.30.0",
+    "pyarrow>=20.0.0",
+    "pydantic>=2.11.5",
+    "pydantic-settings>=2.9.1",
+    "rdata>=0.11.2",
+    "sqlalchemy>=2.0.41",
+]
 
-[tool.poetry.dependencies]
-python = "~3.11"
-dash-bootstrap-components = "^1.4.2"
-dash-daq = "^0.5.0"
-numpy = "^1.26.4"
-pandas = "^2.2.0"
-plotly = "^5.16.1"
-rdata = "^0.10"
-pydantic = "^2.6.1"
-dash = "^2.15.0"
-sqlalchemy = "^2.0.26"
-pydantic-settings = "^2.1.0"
-polars = "^0.20.7"
-pyarrow = "^15.0.0"
-
-[tool.poetry.group.dev.dependencies]
-pytest = "^8.0.2"
-mypy = "^1.8.0"
-pre-commit = "^3.6.2"
-pytest-cov = "^4.1.0"
-ruff = "^0.6"
-pytest-mock = "^3.12.0"
-vulture = "^2.11"
-
-[tool.poetry.group.docs.dependencies]
-pdoc = "^14.4.0"
-
-[tool.poetry.extras]
-docs = ["pdoc"]
-
-[tool.poetry.scripts]
-actigraphy = "actigraphy.__main__:main_entrypoint"
-actigraphy_preprocess = 'actigraphy.__main__:preprocess_entrypoint'
+[dependency-groups]
+dev = [
+    "mypy>=1.16.0",
+    "pre-commit>=4.2.0",
+    "pytest>=8.4.0",
+    "pytest-cov>=6.1.1",
+    "pytest-mock>=3.14.1",
+    "ruff>=0.11.13",
+    "vulture>=2.14",
+]
 
 [build-system]
-requires = ["poetry-core>=1.2.0"]
-build-backend = "poetry.core.masonry.api"
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[project.scripts]
+actigraphy = "actigraphy:app"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/actigraphy"]
 
 [tool.pytest.ini_options]
 pythonpath = [
