### 2025-06-11 030d67f
fix: Correctly call from uv scripts
