+++ pyproject.toml
@@ -35,7 +35,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [project.scripts]
-actigraphy = "actigraphy:app"
+actigraphy = "actigraphy.app:run_app"
 
 [tool.hatch.build.targets.wheel]
 packages = ["src/actigraphy"]
