### 2025-06-11 1ffe0b0
fix: Removed unused docs workflow, update test.yaml to uv
