+++ .github/workflows/test.yaml
@@ -11,19 +11,19 @@ jobs:
     runs-on: ubuntu-latest
     steps:
     - uses: actions/checkout@v4
-    - name: Install poetry
-      run: pipx install poetry
+    - name: Install uv
+      run: pipx install uv
     - uses: actions/setup-python@v5
       with:
         python-version-file: pyproject.toml
-        cache: poetry
+        cache: pip
     - name: Install dependencies
       run: |
-        poetry install
+        uv sync
     - name: Run tests
       id: run-tests
       run: >
-        poetry run pytest \
+        uv run pytest \
           --junitxml=pytest.xml \
           --cov-report=term-missing:skip-covered \
           --cov-report=xml:coverage.xml \
@@ -46,6 +46,6 @@ jobs:
         with:
           python-version: "3.11"
       - run: |
-          pip install poetry
-          poetry install
-          poetry run mypy .
+          pip install uv
+          uv sync
+          uv run mypy .
