### 2025-06-11 eff2e80
Switch Poetry to uv (#101)
* chore: Switch to uv

* fix: Correctly call from uv scripts

* fix: Removed unused docs workflow, update test.yaml to uv
