### 2025-06-26 #8181 @cbrnr
Thank you for your detailed response @nathaliaesper! I tried to run the app, but unfortunately I had trouble with both Docker and uv, which I will explain below.

## Docker
I followed the instructions, but I cannot access the website. Here are the commands I used (the working directory `$PWD` is the Git repository root):

1. `docker pull cmidair/actigraphy:main` (this worked)
2. `git clone https://github.com/childmindresearch/actisleep-tracker.git` (this worked)
4. `docker run -p 127.0.0.1:8051:8051 --volume "$PWD/paper/data/ggir_outputs:/data" cmidair/actigraphy:main`

Here's the output of the final step:

```
   Building actisleep-tracker @ file:///app
      Built actisleep-tracker @ file:///app
Uninstalled 1 package in 0.69ms
Installed 1 package in 0.90ms
2025-06-26 07:59:37,438 - Actigraphy - INFO - Starting Actigraphy app
2025-06-26 07:59:37,447 - Actigraphy - INFO - Attaching callbacks to app
2025-06-26 07:59:37,448 - Actigraphy - INFO - Creating app layout
Dash is running on http://127.0.0.1:8050/

 * Serving Flask app 'Actigraphy'
 * Debug mode: off
WARNING: This is a development server. Do not use it in a production deployment. Use a production WSGI server instead.
 * Running on http://127.0.0.1:8050
Press CTRL+C to quit
```

This looks good to me, but for some reason, the server runs on port 8050 instead of 8051! Navigating to http://127.0.0.1:8050/ opens a website with an error message:

> Internal Server Error
> The server encountered an internal error and was unable to complete your request. Either the server is > overloaded or there is an error in the application.

Two side notes:

1. <kbd>Ctrl</kbd><kbd>C</kbd> does not actually quit, I have to press this key combination at least three times, after which the app finally exits with this message:

```
got 3 SIGTERM/SIGINTs, forcefully exiting
```
2. I think the instructions should not contain the full name of your home directory, please consider using `$PWD` (as I did in my command).

## uv
Running the app with uv gives me a different error.

Here are the step to reproduce (assuming the working directory is still the Git repository root):

1. `uv sync` (this worked)
2. `uv run actigraphy "$PWD/paper/data/ggir_outputs"`

The server starts, but just like with Docker, the URL is http://127.0.0.1:8050/ (so again port 8050 instead of 8051). Navigating to this website now does show me a page, but the images are missing:

![Image](https://github.com/user-attachments/assets/ab99846e-7757-44d0-84ca-13b64f919009)

Clicking on "Load Files" does not do anything, but it gives the following error in the terminal (the example data seems to be in a different format than expected):

```
127.0.0.1 - - [26/Jun/2025 10:11:52] "GET / HTTP/1.1" 200 -
127.0.0.1 - - [26/Jun/2025 10:11:52] "GET /assets/styles.css?m=1750923721.6149025 HTTP/1.1" 200 -
127.0.0.1 - - [26/Jun/2025 10:11:53] "GET /_dash-layout HTTP/1.1" 200 -
127.0.0.1 - - [26/Jun/2025 10:11:53] "GET /_dash-dependencies HTTP/1.1" 200 -
127.0.0.1 - - [26/Jun/2025 10:11:53] "GET /_favicon.ico?v=2.18.2 HTTP/1.1" 200 -
127.0.0.1 - - [26/Jun/2025 10:11:53] "GET /_dash-component-suites/dash/dcc/async-dropdown.js HTTP/1.1" 304 -
127.0.0.1 - - [26/Jun/2025 10:11:53] "GET /assets/CMI_Logo_title.png HTTP/1.1" 200 -
2025-06-26 10:11:57,425 - Actigraphy - INFO - Calling callback: actigraphy.components.file_selection.parse_files.
2025-06-26 10:11:57,426 - Actigraphy - INFO - Creating/loading database
2025-06-26 10:11:57,610 - Actigraphy - ERROR - Subject sub001 not found in database
2025-06-26 10:11:57,610 - Actigraphy - INFO - Subject not found in database. Creating new subject.
/home/clemens/Downloads/actisleep-tracker/.venv/lib/python3.13/site-packages/rdata/parser/_parser.py:1217: UserWarning:

Unknown file type: assumed RDS

/home/clemens/Downloads/actisleep-tracker/.venv/lib/python3.13/site-packages/rdata/parser/_parser.py:1220: UserWarning:

Wrong extension .RData for file in RDS format

2025-06-26 10:11:57,610 - Actigraphy - ERROR - Exception on /_dash-update-component [POST]
Traceback (most recent call last):
  File "/home/clemens/Downloads/actisleep-tracker/src/actigraphy/components/file_selection.py", line 123, in parse_files
    subject = crud.read_subject(session, file_manager["identifier"])
  File "/home/clemens/Downloads/actisleep-tracker/src/actigraphy/database/crud.py", line 31, in read_subject
    raise exceptions.DatabaseError(msg)
actigraphy.core.exceptions.DatabaseError: Subject sub001 not found in database

Traceback (most recent call last):
  File "/home/clemens/Downloads/actisleep-tracker/.venv/lib/python3.13/site-packages/flask/app.py", line 1473, in wsgi_app
    response = self.full_dispatch_request()
  File "/home/clemens/Downloads/actisleep-tracker/.venv/lib/python3.13/site-packages/flask/app.py", line 882, in full_dispatch_request
    rv = self.handle_user_exception(e)
  File "/home/clemens/Downloads/actisleep-tracker/.venv/lib/python3.13/site-packages/flask/app.py", line 880, in full_dispatch_request
    rv = self.dispatch_request()
  File "/home/clemens/Downloads/actisleep-tracker/.venv/lib/python3.13/site-packages/flask/app.py", line 865, in dispatch_request
    return self.ensure_sync(self.view_functions[rule.endpoint])(**view_args)  # type: ignore[no-any-return]
           ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^^^^^^^^^^^^^
  File "/home/clemens/Downloads/actisleep-tracker/.venv/lib/python3.13/site-packages/dash/dash.py", line 1376, in dispatch
    ctx.run(
    ~~~~~~~^
        functools.partial(
        ^^^^^^^^^^^^^^^^^^
    ...<7 lines>...
        )
        ^
    )
    ^
  File "/home/clemens/Downloads/actisleep-tracker/.venv/lib/python3.13/site-packages/dash/_callback.py", line 507, in add_context
    raise err
  File "/home/clemens/Downloads/actisleep-tracker/.venv/lib/python3.13/site-packages/dash/_callback.py", line 496, in add_context
    output_value = _invoke_callback(func, *func_args, **func_kwargs)
  File "/home/clemens/Downloads/actisleep-tracker/.venv/lib/python3.13/site-packages/dash/_callback.py", line 43, in _invoke_callback
    return func(*args, **kwargs)  # %% callback invoked %%
  File "/home/clemens/Downloads/actisleep-tracker/src/actigraphy/core/callback_manager.py", line 93, in wrapper
    return func(*args, **kwargs)
  File "/home/clemens/Downloads/actisleep-tracker/src/actigraphy/components/file_selection.py", line 126, in parse_files
    subject = database_utils.initialize_subject(
        file_manager["identifier"],
    ...<2 lines>...
        session,
    )
  File "/home/clemens/Downloads/actisleep-tracker/src/actigraphy/database/utils.py", line 206, in initialize_subject
    ggir_metadata = ggir_files.MetaData.from_file(ggir_metadata_file)
  File "/home/clemens/Downloads/actisleep-tracker/src/actigraphy/io/ggir_files.py", line 68, in from_file
    metadata = _rdata_to_datadict(filepath)
  File "/home/clemens/Downloads/actisleep-tracker/src/actigraphy/io/ggir_files.py", line 280, in _rdata_to_datadict
    data = rdata.parser.parse_file(filepath)
  File "/home/clemens/Downloads/actisleep-tracker/.venv/lib/python3.13/site-packages/rdata/parser/_parser.py", line 1087, in parse_file
    return parse_data(
        data,
    ...<2 lines>...
        extension=extension,
    )
  File "/home/clemens/Downloads/actisleep-tracker/.venv/lib/python3.13/site-packages/rdata/parser/_parser.py", line 1222, in parse_data
    return parse_function(
        new_data,  # type: ignore [arg-type]
    ...<2 lines>...
        extension=extension,
    )
  File "/home/clemens/Downloads/actisleep-tracker/.venv/lib/python3.13/site-packages/rdata/parser/_parser.py", line 1251, in parse_rdata_binary
    raise NotImplementedError(msg)
NotImplementedError: Unknown file format
127.0.0.1 - - [26/Jun/2025 10:11:57] "POST /_dash-update-component HTTP/1.1" 500 -
```

Side note: You use `https://github.com/childmindresearch/app-actigraphy.git`, but this is just an alias for `https://github.com/childmindresearch/actisleep-tracker.git`. Please only use the latter URL, since this is the public one.

Overall, these issues need to be resolved. In addition, I find the instructions still very difficult to follow, especially in combination with using the example data. I suggest to streamline it even further such that the commands that users need to enter are simple to follow.
