### 2025-07-04 31fd68e
fix: Remove logo in dev version, lock port to 8051.
