### 2025-07-07 956893f
fix: Remove old scaling, change color
