### 2025-07-07 12ce57f
fix: Multiple y-axes [DCH-853] (#107)
* feat: Multiple y-axes

* fix: Remove old scaling, change color
