### 2025-07-16 cd36487
refactor: Disable zoom/pan
