+++ README.md
@@ -53,17 +53,17 @@ Run ActiSleep Tracker through Docker by typing:
 ```bash
    docker run \
    -p 127.0.0.1:8051:8051 \
-   --volume /Users/nathalia.esper/Documents/actisleep-tracker/paper/data/ggir_outputs:/data \
+   --volume $PWD/actisleep-tracker/paper/data/ggir_outputs:/data \
    cmidair/actigraphy:main
 ```
-If you are using your own dataset, replace `/Users/nathalia.esper/Documents/actisleep-tracker/paper/data/ggir_outputs` with the path to the GGIR's outputs.
+If you are using your own dataset, replace `$PWD/actisleep-tracker/paper/data/ggir_outputs` with the path to the GGIR's outputs.
 
 The application will the available at http://localhost:8051.
 
 Select the output you would like to visualize; then, click on `Load Files`:
 
 <figure>
-  <img src="paper/man/README-initialscreen.png" alt="Alt text" width="100%"/>
+  <img src="paper/man/README-initialscreen2.png" alt="Alt text" width="100%"/>
   <figcaption>Figure 1: ActiSleep Tracker initial screen. To start, select the data you would like to load and click `Load Files`.</figcaption>
 </figure>
 
@@ -72,23 +72,24 @@ Note: the file may take some time to load for the first time.
 Once the file is loaded, you will see the following screen:
 
 <figure>
-  <img src="paper/man/README-loaddata.png" alt="Alt text" width="100%"/>
-  <figcaption>Figure 2: Main screen. From here, you can change the sleep annotations from your data and add multiple slee windows.</figcaption>
+  <img src="paper/man/README-loaddata2.png" alt="Alt text" width="100%"/>
+  <figcaption>Figure 2: Main screen. From here, you can change the sleep annotations from your data and add multiple sleep windows.</figcaption>
 </figure>
+
 
 The top slider is used to select the day to visualize. Next, you have three main questions to answer about that night:
 1. Are there multiple sleep periods in these data? - Select this toggle button if you identify more than one sleep window on that day (i.e., multiple sleep windows at night or nap times during the day).
 2. Do you need to review this night? - Select this toggle button if that day will need further review.
 3. Are >2 hours of data missing between 8PM to 8AM? - Select this toggle button if there is insufficient data for sleep analysis. For the protocol we designed, we were considering sleep periods between 8PM to 8AM, but you can use this toggle and adapt to your own needs.
 
-The graph represents the angle of sensor's z-axis (in blue) and arm movement (in black). The slider below the graph is used to adjust the main sleep window. To change the window, just slide the slider to the desired onset and/or offset sleep times. In cases where your data has more than one sleep window (multiple sleep windows at night or nap times during the day), you can use the `Add Slider` button to create a new sleep window. Note that the main sleep episode should always use the slider below the graphs. The code will automatically save all changes you made and will create three files:
+The graph represents the angle of sensor's z-axis (in blue) and arm movement (in grey). The slider below the graph is used to adjust the main sleep window. To change the window, just slide the slider to the desired onset and/or offset sleep times. In cases where your data has more than one sleep window (multiple sleep windows at night or nap times during the day), you can use the `Add Slider` button to create a new sleep window. Note that the main sleep episode should always use the slider below the graphs. The code will automatically save all changes you made and will create three files:
 
[…]
