+++ README.md
@@ -51,12 +51,13 @@ where `outputs` contains all files generated by GGIR; `raw` is the raw actigraph
 Run ActiSleep Tracker through Docker by typing:
 
 ```bash
+   DATA_DIR="INSERT_PATH_TO_YOUR_DATA_DIRECTORY"
    docker run \
    -p 127.0.0.1:8051:8051 \
    --volume $DATA_DIR:/data \
    cmidair/actigraphy:main
 ```
-If you are using your own dataset, replace `$DATA_DIR` with the path to the GGIR's outputs.
+If you want to try our sample dataset, navigate to the root of the repository and use `DATA_DIR=./paper/data/ggir_outputs`.
 
 The application will the available at http://localhost:8051.
 
