### 2025-07-18 90c1156
fix: Faulty host in Docker image
