### 2025-07-18 21b4165
fix: Attach health endpoint outside of creation to avoid mocking it.
