+++ src/actigraphy/app.py
@@ -47,14 +47,20 @@ def create_app() -> dash.Dash:
         ),
     )
 
-    @app.server.route("/health")
-    def health() -> str:
-        return "{status: ok}"
-
     return app
+
+
+def attach_health_endpoint(app: dash.Dash) -> None:
+    """Attaches the health endpoint to the app."""
+
+    @app.server.route("/health")  # type: ignore[misc] # App is untyped.
+    def health() -> str:
+        """Returns only a status: 200."""
+        return ""
 
 
 def run_app() -> None:
     """Entrypoint for the application."""
     app = create_app()
+    attach_health_endpoint(app)
     app.run(debug=False, host="0.0.0.0", port=8051)  # noqa: S104
