+++ README.md
@@ -20,10 +20,10 @@ The application may be installed either through Docker (recommended for users) o
    ```
 3. Run the Docker image with your data:
    ```bash
-   docker run \
-      -p 127.0.0.1:8051:8051 \
-      --volume /local/path/to/data:/data \
-      cmidair/actigraphy:main
+   docker run
+       -p 127.0.0.1:8051:8051
+       --volume /local/path/to/data:/data
+       cmidair/actigraphy:main
    ```
 Replace `/local/path/to/data` with the path to your input data directory (i.e., GGIR's output folder).
 
@@ -53,7 +53,7 @@ Run ActiSleep Tracker through Docker by typing:
 ```bash
    DATA_DIR="INSERT_PATH_TO_YOUR_DATA_DIRECTORY"
    docker run \
-   -p 8051:8051 \
+   -p 127.0.0.1:8051:8051 \
    --volume $DATA_DIR:/data \
    cmidair/actigraphy:main
 ```
