### 2025-07-18 f716fec
fix: Faulty host in Docker image (#111)
* fix: Faulty host in Docker image

* fix: Attach health endpoint outside of creation to avoid mocking it.

* update readme

* fix: Backslashes
