### 2025-07-18 fd831bb
Update dash requirement
Updates the requirements on [dash](https://github.com/plotly/dash) to permit the latest version.

Updates `dash` to 3.1.1
- [Release notes](https://github.com/plotly/dash/releases)
- [Changelog](https://github.com/plotly/dash/blob/dev/CHANGELOG.md)
- [Commits](https://github.com/plotly/dash/compare/v2.15.0...v3.1.1)

---
updated-dependencies:
- dependency-name: dash
  dependency-version: 3.1.1
  dependency-type: direct:production
  dependency-group: production-dependencies
...

Signed-off-by: dependabot[bot] <support@github.com>
