+++ pyproject.toml
@@ -5,7 +5,7 @@ description = "Add your description here"
 readme = "README.md"
 requires-python = "~=3.11.10"
 dependencies = [
-    "dash~=2.15",
+    "dash~=3.1",
     "dash-bootstrap-components>=1.7.1",
     "dash-daq>=0.6.0",
     "numpy>=2.3.0",
