### 2025-07-18 fef2c7c
fix: Using a relative path via UV results in an error finding files
