### 2025-07-23 3af928a
Update README with Git LFS, fix local relative paths (#112)
* fix: Using a relative path via UV results in an error finding files

* doc: Add explicit Git LFS instruction

* fix: Fix CLI getsubject directories test
