### 2025-07-24 #8181 @nathaliaesper
Hi @richrobe and @cbrnr,

Thank you for your comments. We have completed adjusting the application to incorporate your comments and suggestions and resolved the issues you experienced when starting the application using Docker and UV. Please see below our comments on each one of the issues:

From @richrobe:

1. Double-clicking to reset the original view causes a misalignment between the plot and the sliders. Maybe this can be fixed by redrawing the plot after resetting the view? Adjusting the sliders after resetting the view re-aligns the plot.
_Indeed, a misalignment problem occurred between the plot and the sliders when the zooming tool was used. Unfortunately, this appears to be a problem in Plotly Python, and we were unable to override the double-click behavior. To fix this issue, we disabled the zooming function on our plots._

2. What is the y-axis of the plot? And since I assume that the units for x-axis angle and arm movement are not the same, I would suggest splitting the y-axis into two axes (one left of the plot and one right of the plot) to make it clearer what is plotted on which axis.
_Thank you for noticing this. We adjusted the plot to include units on the y-axis for the x-axis angle (on the left side of the plot) and arm movement (on the right side of the plot), along with their corresponding unit measures._

From @cbrnr:

1. Internal Server Error message when running through Docker.
_Recently, the default host was changed on Dash. We have resolved the issue, and the application is now operational on the correct host and port._

2. Header figures on the website were not being displayed when running through uv.
We decided to remove the header images from the website, as they do not provide any additional information to the plots._

3. Clicking on "Load Files" does not do anything, but gives an error in the terminal (the example data seems to be in a different format than expected).
_The files for the example data were added to GitHub using Git LFS. To run the application with the example data, you will need to run the following commands:_
`git clone https://github.com/childmindresearch/actisleep-tracker.git
cd actisleep-tracker
git lfs install
git lfs pull
docker run -p 127.0.0.1:8051:8051 --volume $DATA_DIR:/data cmidair/actigraphy:main`
_We changed the instructions on the main page to reflect these alterations._

4. Change URL on the README file
_Thank you for bringing that to our attention. We changed the placeholder URL to the correct one (https://github.com/childmindresearch/actisleep-tracker.git)_

5. Instructions are difficult to follow
_We made a few changes to the instructions to make it easier to understand and run the application_
