### 2025-07-25 #8181 @cbrnr
Thank you! I confirm that now everything works with uv! Just a small note, the `README.md` still says:

> The application may be installed either through Docker (recommended for users) or Poetry (recommended for developers).

Please replace Poetry with uv.

A small suggestion (not a blocker though), it is a bit confusing that I can add a second slider an *not* enable the "Are there multiple sleep periods in these data?" toggle. Ideally, the "Add Slider" button should be grayed out unless that toggle is enabled. But again, no blocker.

Finally, I cannot try the Docker image, so it would be great if @richrobe confirmed that the new build is working. Other than that, I am now happy with the instructions and the other things I've mentioned in my previous comment.
