### 2025-07-25 #8181 @richrobe
Yes, the Docker image works for me (see screenshot attached for confirmation). However, one additional thing arose that might be helpful for future users:
At first, it didn't work as I forgot to correctly specify the mounting of the data folder, (i.e., I forgot to specify the second part of the mapping `/local/path/to/data:/data`). The error I received was the following:

```
Traceback (most recent call last):
  File "/app/.venv/bin/actigraphy", line 10, in <module>
    sys.exit(run_app())
             ^^^^^^^^^
  File "/app/src/actigraphy/app.py", line 64, in run_app
    app = create_app()
          ^^^^^^^^^^^^
  File "/app/src/actigraphy/app.py", line 44, in create_app
    file_selection.file_selection(subject_directories),
    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/app/src/actigraphy/components/file_selection.py", line 46, in file_selection
    dropdown_choices[0],
    ~~~~~~~~~~~~~~~~^^^
IndexError: list index out of range
```

As I can imagine that this might also happen to other potential users, I'd suggest making users more aware of this by issuing a more informative error message (maybe something like "No files found. Did you check if you correctly mounted the volume into the docker container?") and/or also give an additional hint to users in the README to double-check the correct data path?

Other than that, I'm also quite happy with the package!
