### 2025-07-25 #8181 @nathaliaesper
Hi @cbrnr,
Yes, I updated the article. Please, see the latest version on our GitHub repository (https://github.com/childmindresearch/actisleep-tracker/tree/main/paper).

I will review the automated tests and community guidelines you mentioned.
