### 2025-07-29 #8181 @cbrnr
I have now read the manuscript and I only have two minor comments:

1. If there is a need to shorten the manuscript, I think Figure 2 (A) could be removed, as panel (B) is sufficient to showcase the functionality of the package. That way, the manuscript would be half a page shorter.
2. Page 2, line 51: "This ensured..." -> "This ensures..."

Otherwise, I think this is a great summary and ready for publication.
