### 2025-08-04 #8181 @juliaferraioli
Hi  @nathaliaesper, once you have removed the figure, could you please:

- List the DOI of the archived version here.

I can then move forward with recommending acceptance of the submission. Thanks!
