### 2025-08-05 1cb888e
Figures were changed to reflect the latest version of the software
