### 2025-08-05 #8181 @nathaliaesper
Hi @juliaferraioli,

- We changed Figure 2 as suggested by the reviewers and updated the text to follow the new Figure 2.
- We made a tagged release of our software. The release version is v1.0.0
- We archived the reviewed software in Zenodo and made sure that the archival deposit has the correct metadata.
- Here is the DOI for the archived version: 10.5281/zenodo.16748475

Let me know if you need anything else from me.
