### 2025-08-08 #8181 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
