### 2025-08-11 423d0fc
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/astral-sh/ruff-pre-commit: v0.12.2 → v0.12.8](https://github.com/astral-sh/ruff-pre-commit/compare/v0.12.2...v0.12.8)
- [github.com/pre-commit/pre-commit-hooks: v4.5.0 → v6.0.0](https://github.com/pre-commit/pre-commit-hooks/compare/v4.5.0...v6.0.0)
